The discojuice feeds are downloading by the CRON job.
