+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/ClarinDiscoJuiceFeedsControllerIT.java
@@ -32,6 +32,7 @@ public class ClarinDiscoJuiceFeedsControllerIT extends AbstractControllerIntegra
         String authTokenAdmin = getAuthToken(eperson.getEmail(), password);
 
         // Expected response created from the test file: `discofeedResponse.json`
+        // Wrapped to the `callback` string = `dj_md_1`
         String responseString = "dj_md_1([{\"country\":\"CZ\",\"keywords\":[\"Identity Provider for employees and " +
                 "readers of the Archiepiscopal Gymnasium in Kromeriz - Library\",\"Identity Provider pro" +
                 " zamÄ\\u203Astnance a ÄŤtenĂˇĹ™e knihovny ArcibiskupskĂ©ho gymnĂˇzia v KromÄ\\u203AĹ™Ă\u00ADĹľi\"," +
@@ -49,6 +50,7 @@ public class ClarinDiscoJuiceFeedsControllerIT extends AbstractControllerIntegra
                 "\"The Research Library in Hradec KrĂˇlovĂ©\"}])";
 
         // Load bitstream from the item.
+        // Request with callback
         getClient(authTokenAdmin).perform(get("/api/discojuice/feeds?callback=dj_md_1"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(APPLICATION_JAVASCRIPT_UTF8))
@@ -60,9 +62,29 @@ public class ClarinDiscoJuiceFeedsControllerIT extends AbstractControllerIntegra
     public void shouldReturnNoContent() throws Exception {
         String authTokenAdmin = getAuthToken(eperson.getEmail(), password);
 
+        // Not wrapped to the callback string
+        String responseString = "[{\"country\":\"CZ\",\"keywords\":[\"Identity Provider for employees and " +
+                "readers of the Archiepiscopal Gymnasium in Kromeriz - Library\",\"Identity Provider pro" +
+                " zamÄ\\u203Astnance a ÄŤtenĂˇĹ™e knihovny ArcibiskupskĂ©ho gymnĂˇzia v KromÄ\\u203AĹ™Ă\u00ADĹľi\"," +
+                "\"ArcibiskupskĂ© gymnĂˇzium v KromÄ\\u203AĹ™Ă\u00ADĹľi - Knihovna\"],\"entityID\":\"https:" +
+                "\\/\\/agkm.cz\\/idp\\/shibboleth\",\"title\":\"Archiepiscopal Gymnasium in Kromeriz - Library\"}," +
+                "{\"country\":\"CZ\",\"keywords\":[\"Identity Provider for staff of the Institute of Agricultural " +
+                "Economics and Information and patrons of the AntonĂ\u00ADn Ĺ vehla Library\",\"Identity Provider" +
+                " pro zamÄ\\u203Astnance ĂšZEI a ÄŤtenĂˇĹ™e Knihovny AntonĂ\u00ADna Ĺ vehly\",\"Ăšstav " +
+                "zemÄ\\u203AdÄ\\u203AlskĂ© ekonomiky a informacĂ\u00AD\"],\"entityID\":\"https:\\/\\/aleph" +
+                ".uzei.cz\\/idp\\/shibboleth\",\"title\":\"Institute of Agricultural Economics and Information\"}," +
+                "{\"country\":\"CZ\",\"keywords\":[\"Identity Provider for patrons and staff of the Research Library" +
+                " in Hradec KrĂˇlovĂ©\",\"Identity Provider pro ÄŤtenĂˇĹ™e a zamÄ\\u203Astance StudijnĂ\u00AD a" +
+                " vÄ\\u203AdeckĂ© knihovny v Hradci KrĂˇlovĂ©\",\"StudijnĂ\u00AD a vÄ\\u203AdeckĂˇ knihovna v" +
+                " Hradci KrĂˇlovĂ©\"],\"entityID\":\"https:\\/\\/aleph.svkhk.cz\\/idp\\/shibboleth\",\"title\":" +
+                "\"The Research Library in Hradec KrĂˇlovĂ©\"}]";
+
         configurationService.setProperty("shibboleth.discofeed.url", "non existing endpoint");
         // Load bitstream from the item.
-        getClient(authTokenAdmin).perform(get("/api/discojuice/feeds?callback=dj_md_1"))
-                .andExpect(status().isNoContent());
+        // Request without callback
+        getClient(authTokenAdmin).perform(get("/api/discojuice/feeds"))
+                .andExpect(status().isOk())
+                .andExpect(content().contentType(APPLICATION_JAVASCRIPT_UTF8))
+                .andExpect(content().string(responseString));
     }
 }
