+++ dspace/config/clarin-dspace.cfg
@@ -78,8 +78,8 @@ shibboleth.discofeed.url = https://lindat.mff.cuni.cz/Shibboleth.sso/DiscoFeed
 
 # Configuration properties for DiscoJuice
 discojuice.feeds = edugain, dfn, cesnet, surfnet2, haka, kalmar
-# refresh time in hours
-discojuice.refresh = 2
+# CRON job refresh time definition - default is refresh in every 2 hours.
+discojuice.refresh = 0 */2 * * *
 # Comma separated list of entityIDs; we try to guess country on these
 discojuice.rewriteCountries = https://idp.scc.kit.edu/idp/shibboleth, https://fedauth.london.edu/oala/metadata, https://youidlite.youid.net/idp/shibboleth, https://cavle.org/shibboleth
